requirements.txt
python-telegram-bot>=21.0
telethon>=1.36
sqlalchemy>=2.0
aiosqlite>=0.19
python-dotenv>=1.0